refactor(openomni): merge ingress handler stage files into handlers.ts
The dispatch lifecycle was chopped into 7 files (handlers, -events,
-writeback, -worker-run, -worker-dispatch, -prompt, -types; 495 LOC)
with a 7-field HandlerContext type threaded through 5 stage files and
handler-worker-dispatch importing 6 siblings for its single importer.
Fold the stages back into one handlers.ts: bodies are byte-identical
(only the commitWriteback import alias returns to its real name and
the finishCoordinatorDispatch argument list re-wraps), the stage
helpers become namespace-local, and the HandlerContext indirection
file dies. No behavior change; external surface stays
IngressHandlers.{HandlerContext,buildExecutionRequest,handleResident,handleDirect}
(the unused namespace dispatchWritebackCommit re-export wrapper is
dropped — zero importers).
